% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggs.R
\name{aggs}
\alias{aggs}
\title{elastic DSL aggs}
\usage{
aggs(.obj = list(), ...)
}
\arguments{
\item{.obj}{An index object. If nothing passed defaults to all indices, equivalent to
doing e.g., \code{localhost:9200/_search}}

\item{...}{Further args passed on}
}
\description{
elastic DSL aggs

Aggregations
}
\examples{
library(elastic)
target <- Search(
index = "gbif",
body = list(aggs = list(statistic = list(avg = list(field = "decimalLatitude")))))

aggs_example <- index("gbif") \%>\%
  aggs(x = list(statistic = list(avg = list(field = "decimalLatitude"))))

identical(target$aggregations, aggs_example$aggregations)

}

