% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bool.R
\name{bool}
\alias{bool}
\alias{bool_}
\title{bool dsl}
\usage{
bool(.obj = list(), ...)

bool_(.obj = list(), ..., .dots)
}
\arguments{
\item{.obj}{An index object. If nothing passed defaults to all indices, equivalent to
doing e.g., \code{localhost:9200/_search}}

\item{...}{Further args passed on}

\item{.dots}{Explanation...}
}
\description{
bool dsl
}
\examples{
\dontrun{
elastic::connect(errors = "complete")

bool(must_not = list(term=list(speaker="KING HENRY IV")))
index("shakespeare") \%>\%
   bool(must_not = list(term=list(speaker="KING HENRY IV")))
index("shakespeare") \%>\%
   bool(should = list(list(term=list(speech_number=1))))
}
}

