% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boosting.R
\name{boosting}
\alias{boosting}
\alias{boosting_}
\title{boosting dsl}
\usage{
boosting(.obj = list(), ..., negative_boost = NULL)

boosting_(.obj = list(), ..., .dots, negative_boost = NULL)
}
\arguments{
\item{.obj}{An index object. If nothing passed defaults to all indices, equivalent to
doing e.g., \code{localhost:9200/_search}}

\item{...}{Further args passed on}

\item{negative_boost}{Explanation...}

\item{.dots}{Explanation...}
}
\description{
boosting dsl
}
\examples{
\dontrun{
elastic::connect(errors = "complete")

# boosting query
boost <- '{
 "query" : {
  "boosting" : {
      "positive" : {
          "term" : {
              "play_name" : "henry"
          }
      },
      "negative" : {
          "term" : {
              "text_entry" : "thou"
          }
      },
      "negative_boost" : 0.8
    }
 }
}'
elastic::Search(index="shakespeare", body=boost)
index("shakespeare") \%>\%
 boosting(positive = list(term = list(play_name = "henry")),
   negative_boost = 0.8)
index("shakespeare") \%>\%
 boosting(positive = list(term = list(play_name = "henry")),
         negative = list(term = list(text_entry = "thou")),
         negative_boost = 0.8)

}
}

