% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{common}
\alias{common}
\alias{common_}
\title{common dsl}
\usage{
common(
  .obj = list(),
  field,
  query = NULL,
  cutoff_frequency = NULL,
  low_freq_operator = NULL,
  minimum_should_match = NULL
)

common_(
  .obj = list(),
  field,
  query = NULL,
  cutoff_frequency = NULL,
  low_freq_operator = NULL,
  minimum_should_match = NULL
)
}
\arguments{
\item{.obj}{An index object. If nothing passed defaults to all indices, equivalent to
doing e.g., \code{localhost:9200/_search}}

\item{field}{Explanation...}

\item{query}{Explanation...}

\item{cutoff_frequency}{Explanation...}

\item{low_freq_operator}{Explanation...}

\item{minimum_should_match}{Explanation...}
}
\description{
common dsl
}
\examples{
\dontrun{
elastic::connect(errors = "complete")

# common query - not working yet
# index("shakespeare") \%>\% common( speech_number <= 5 )

## common query without DSL
x <- '{
  "query": {
    "query_string" : {
      "default_field" : "text_entry",
      "query" : "this AND that OR thus"
    }
  }
}'
Search("shakespeare", body = body)

body <- '{
 "query" : {
   "common": {
      "body": {
           "query": "against"
       }
     }
  }
}'
elastic::Search('shakespeare', body=body)
}
}
