% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Explain a query}
\usage{
describe(.data)
}
\arguments{
\item{.data}{(list) input, using higher level interface}
}
\description{
Explain a query
}
\examples{
elastic::connect()

index("shakespeare") \%>\% range( speech_number <= 5 ) \%>\% describe

index("shakespeare") \%>\%
   bool(must_not = list(term=list(speaker="KING HENRY IV"))) \%>\%
   describe

index("geoshape") \%>\%
   geoshape(field = "location", type = "envelope",
            coordinates = list(c(-30, 50), c(30, 0))) \%>\%
   describe()
}

