% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dont_exec.R
\name{dont_exec}
\alias{dont_exec}
\title{Prevent executing search}
\usage{
dont_exec(.obj)
}
\arguments{
\item{.obj}{An index object. If nothing passed defaults to all indices, equivalent to
doing e.g., \code{localhost:9200/_search}}
}
\description{
Prevent executing search
}

