% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\name{fields}
\alias{fields}
\alias{fields_}
\title{Retrieve specific fields}
\usage{
fields(.data = list(), ...)

fields_(.data = list(), ..., .dots)
}
\arguments{
\item{.data}{Input. An object of class \code{esdsl}, or something
that can be coerced to that}

\item{...}{Comma separated list of unquoted variable names}

\item{.dots}{Used to work around non-standard evaluation}
}
\description{
Retrieve specific fields
}
\examples{
\dontrun{
index("shakespeare") \%>\% fields(play_name, speaker)
}
}
