% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter}
\alias{filter}
\title{elastic DSL filters}
\usage{
filter(x)
}
\arguments{
\item{.obj}{An index object. If nothing passed defaults to all indices, equivalent to
doing e.g., \code{localhost:9200/_search}}

\item{...}{xxxx}
}
\description{
new setup, not quite ready yet
this will require queries to be passed inside this function call
to indicate that you want a query to be a filter type query
}
\examples{
\dontrun{
filter(term, speaker = "KING HENRY IV")
}
}
\keyword{internal}
