% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hits.R
\name{hits}
\alias{hits}
\title{Get results (hits slot)}
\usage{
hits(x)
}
\arguments{
\item{x}{Input}
}
\description{
Get results (hits slot)
}
\examples{
\dontrun{
index("shakespeare") \%>\%
 ids(1, 2, 150) \%>\%
 hits()
}
}
