% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{ids}
\alias{ids}
\alias{ids_}
\title{ids dsl}
\usage{
ids(.obj = list(), ..., type = NULL)

ids_(.obj = list(), ..., .dots, type = NULL)
}
\arguments{
\item{.obj}{An index object. If nothing passed defaults to all indices, equivalent to
doing e.g., \code{localhost:9200/_search}}

\item{...}{Further args passed on}

\item{type}{type of id to match on. Default: \code{NULL}}

\item{.dots}{Explanation...}
}
\description{
ids dsl
}
\examples{
\dontrun{
elastic::connect(errors = "complete")

# ids query
index("shakespeare") \%>\%
 ids(1, 2, 150) \%>\%
 n()
}
}
