% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/index.R
\name{index}
\alias{index}
\title{Set index, and optionally type}
\usage{
index(index, type = NULL, what = "mappings", ...)
}
\arguments{
\item{index}{Index name}

\item{type}{Type name, Default: NULL, so all types}

\item{what}{In \code{\link{index}}, whether to get mappings, aliases, or settings.}

\item{...}{Further args passed on}
}
\description{
Set index, and optionally type
}
\examples{
\dontrun{
index("shakespeare")
index("shakespeare", "scene")
index("shakespeare", "act")
index("plos")
index("gbif")
}
}

