% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_all.R
\name{match_all}
\alias{match_all}
\title{match_all dsl}
\usage{
match_all(.obj = list(), boost = NULL)
}
\arguments{
\item{.obj}{An index object. If nothing passed defaults to all indices, equivalent to
doing e.g., \code{localhost:9200/_search}}

\item{boost}{Explanation...}
}
\description{
match_all dsl
}
\examples{
\dontrun{
elastic::connect(errors = "complete")

# match all
index("shakespeare") \%>\% match_all()
index("shakespeare") \%>\% match_all(boost = 2)
}
}
