% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n.R
\name{n}
\alias{n}
\title{Get a count of number of documents found}
\usage{
n(.data)
}
\arguments{
\item{.data}{Input}
}
\description{
Get a count of number of documents found
}
\examples{
\dontrun{
index("gbif") \%>\% n()
}
}

