% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefix.R
\name{prefix}
\alias{prefix}
\alias{prefix_}
\title{prefix dsl}
\usage{
prefix(.obj = list(), ...)

prefix_(.obj = list(), ..., .dots)
}
\arguments{
\item{.obj}{An index object. If nothing passed defaults to all indices, equivalent to
doing e.g., \code{localhost:9200/_search}}

\item{...}{Further args passed on}

\item{.dots}{Explanation...}
}
\description{
prefix dsl
}
\examples{
\dontrun{
elastic::connect(errors = "complete")

# prefix query
index("shakespeare") \%>\%
 prefix(speaker = "we") \%>\%
 n()
}
}

