% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\alias{query_}
\title{Elasticsearch DSL}
\usage{
query(.obj = list(), ...)

query_(.obj = list(), ..., .dots)
}
\arguments{
\item{.obj}{An index object. If nothing passed defaults to all indices, equivalent to
doing e.g., \code{localhost:9200/_search}}

\item{...}{Further args passed on}

\item{.dots}{Explanation...}
}
\description{
Elasticsearch DSL
}
\details{
The DSL for \code{elastic} makes it easy to do queries against an Elasticsearch
instance, either local or remote.

The workflow with the DSL:
\enumerate{
 \item Start with the index to use, e.g., \code{index("shakespeare")}
 \item Define queries, e.g., \code{range(speech_number == 5, line_id > 3)}
 \item Execute search, e.g., \code{Search()}
}

Alternatively, if nothing follows a query definition, \code{\link{Search}} is called
to execute the search with the query as given. In a sense, this is essentially like
what \code{dplyr} does.
}
\examples{
\dontrun{
elastic::connect(errors = "complete")

query()
}
}

