% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.R
\name{range}
\alias{range}
\alias{range_}
\title{range dsl}
\usage{
range(.obj = list(), ..., boost = 1, time_zone = NULL, execution = NULL,
  cache = FALSE)

range_(.obj = list(), ..., .dots)
}
\arguments{
\item{.obj}{An index object. If nothing passed defaults to all indices, equivalent to
doing e.g., \code{localhost:9200/_search}}

\item{...}{Further args passed on}

\item{boost}{Explanation...}

\item{time_zone}{Explanation...}

\item{execution}{Explanation...}

\item{cache}{Explanation...}

\item{.dots}{Explanation...}
}
\description{
range dsl
}
\examples{
\dontrun{
elastic::connect(errors = "complete")

x <- index("shakespeare") \%>\% range( speech_number <= 5 ) \%>\% size(200)
max(vapply(x$hits$hits, "[[", 1, c("_source", "speech_number")))
x <- index("shakespeare") \%>\% range( speech_number >= 5 ) \%>\% size(200)
min(vapply(x$hits$hits, "[[", 1, c("_source", "speech_number")))
# index("shakespeare") \%>\% range( speech_number <= c(1,5) ) # doens't work
# index("shakespeare") \%>\% range( speech_number >= c(1,5) ) # doens't work
}
}

