% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/search.R
\name{Search_}
\alias{Search_}
\title{Search method for the DSL}
\usage{
Search_(.obj = "", ...)
}
\arguments{
\item{.obj}{things stuff}

\item{...}{Further args passed on to \code{\link[elastic]{Search}}}
}
\description{
Search method for the DSL
}

