% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.R
\name{sort}
\alias{sort}
\alias{sort_}
\title{Sort documents}
\usage{
sort(.data = list(), ...)

sort_(.data = list(), ..., .dots)
}
\arguments{
\item{.data}{Input. An object of class \code{esdsl}, or something
that can be coerced to that}

\item{...}{Comma separated list of unquoted variable names}

\item{.dots}{Used to work around non-standard evaluation}
}
\description{
Sort documents
}
\examples{
\dontrun{
index("gbif") \%>\% sort(asc(county))
index("gbif") \%>\% sort(asc(county), desc(country))
}
}
