% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tabl.R
\name{tabl}
\alias{tabl}
\title{Attempt to coerce output to a single data.frame}
\usage{
tabl(x, source_only = TRUE, n = 10)
}
\arguments{
\item{x}{Input}

\item{source_only}{(logical) Return source only?}

\item{n}{(integer) Count to print.}
}
\description{
Attempt to coerce output to a single data.frame
}
\details{
Note that this may not always succeed, as we can't predict what
various ways your data in formatted in your Elasticsearch instance.
}
\examples{
\dontrun{
elastic::connect()
x <- index("shakespeare") \%>\%
  filter() \%>\%
  prefix(speaker = "we") \%>\%
  size(200) \%>\%
  exec()
x \%>\% tabl()

x <- index("shakespeare") \%>\%
  filter() \%>\%
  range(line_number <= 20) \%>\%
  size(200) \%>\%
  exec
x \%>\% tabl()
}
}

