% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term.R
\name{term}
\alias{term}
\alias{term_}
\title{term dsl}
\usage{
term(.obj = list(), ...)

term_(.obj = list(), ..., .dots)
}
\arguments{
\item{.obj}{An index object. If nothing passed defaults to all indices, equivalent to
doing e.g., \code{localhost:9200/_search}}

\item{...}{Further args passed on}

\item{.dots}{Explanation...}
}
\description{
term dsl
}
\examples{
\dontrun{
elastic::connect(errors = "complete")

index("shakespeare") \%>\% term(speaker = HOTSPUR)
}
}

