% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggs.R
\name{aggs}
\alias{aggs}
\title{elastic DSL aggs}
\usage{
aggs(.obj = list(), ...)
}
\arguments{
\item{.obj}{An index object. If nothing passed defaults to all indices, equivalent to
doing e.g., \code{localhost:9200/_search}}

\item{...}{Further args passed on}
}
\description{
elastic DSL aggs

Aggregations
}

