% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.query.R
\name{as.query}
\alias{as.query}
\title{As query functions}
\usage{
as.query(x, comb = FALSE, ...)
}
\arguments{
\item{x}{Input}

\item{comb}{(logical) A combined query or not}

\item{...}{Ignored}
}
\description{
As query functions
}
\keyword{internal}

