% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/elasticdsl-package.r
\docType{package}
\name{elasticdsl}
\alias{elasticdsl}
\alias{elasticdsl-package}
\title{elasticdsl: An Elasticsearch DSL R client.}
\description{
elasticdsl: An Elasticsearch DSL R client.
}

