% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/filter.R
\name{filters}
\alias{and}
\alias{bool}
\alias{bool_}
\alias{filter}
\alias{filters}
\alias{ids}
\alias{ids_}
\alias{not}
\alias{or}
\alias{prefix}
\alias{prefix_}
\alias{range}
\alias{range_}
\title{elastic DSL filters}
\usage{
filter(x)

or(x)

and(x)

not(x)

range(.obj = list(), ..., boost = 1, time_zone = NULL, execution = NULL,
  cache = FALSE)

range_(.obj = list(), ..., .dots)

bool(.obj = list(), ...)

bool_(.obj = list(), ..., .dots)

prefix(.obj = list(), ...)

prefix_(.obj = list(), ..., .dots)

ids(.obj = list(), ...)

ids_(.obj = list(), ..., .dots)
}
\arguments{
\item{x}{Input to various functions}

\item{.obj}{An index object. If nothing passed defaults to all indices, equivalent to
doing e.g., \code{localhost:9200/_search}}

\item{...}{Further args passed on}

\item{index}{Index name}

\item{type}{Type name, Default: NULL, so all types}
}
\description{
elastic DSL filters
}
\examples{
\dontrun{
# DSL filters default to search across all indices
bool(must_not = list(term=list(speaker="KING HENRY IV")))

# At the end either
## explain
index("shakespeare") \%>\%
 ids(c(1, 2, 150)) \%>\%
 explain()
## or execute
index("shakespeare") \%>\%
 ids(c(1, 2, 150)) \%>\%
 exec()

# boolean operators, use to assign within the filter
index("shakespeare") \%>\% and() \%>\% attributes
index("shakespeare") \%>\% or() \%>\% attributes
index("shakespeare") \%>\% not() \%>\% attributes

# prefix filter
index("shakespeare") \%>\%
 prefix(speaker = "we") \%>\%
 n()

# ids filter
index("shakespeare") \%>\%
 ids(c(1, 2, 150)) \%>\%
 n()

# chain queries/filters together
index("shakespeare") \%>\%
 filter() \%>\%
 or() \%>\%
 prefix(speaker = "we") \%>\%
 ids(c(1, 2, 150)) \%>\%
 exec() \%>\%
 n()

# not filter
index("shakespeare") \%>\%
 filter() \%>\%
 not() \%>\%
 prefix(speaker = "we") \%>\%
 exec()

# bool filter
index("shakespeare") \%>\%
   bool(must_not = list(term=list(speaker="KING HENRY IV"))) \%>\%
   exec
index("shakespeare") \%>\%
   bool(should = list(list(term=list(speech_number=1)))) \%>\%
   exec

# range filter
index("shakespeare") \%>\%
   range( speech_number <= 5 )
index("gbif") \%>\%
   filter() \%>\%
   range( decimalLatitude <= 40 ) \%>\%
   exec()
index("gbif") \%>\%
   filter() \%>\%
   or() \%>\%
   range(decimalLatitude <= 20) \%>\%
   range(decimalLongitude <= 10) \%>\%
   exec()
}
}

