% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hits.R
\name{hits}
\alias{hits}
\title{Get results}
\usage{
hits(x)
}
\arguments{
\item{x}{Input}
}
\description{
Get results
}
\examples{
\dontrun{
index("shakespeare") \%>\%
 ids(c(1, 2, 150)) \%>\%
 exec() \%>\%
 hits()
}
}

