% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.R
\name{size}
\alias{from}
\alias{size}
\title{Specify how many documents to return}
\usage{
size(.data, size)

from(.data, from)
}
\arguments{
\item{.data}{Input}

\item{size}{(numeric) number of documents to return}
}
\description{
Specify how many documents to return
}
\examples{
\dontrun{
index("gbif") \%>\% size(1)
index("gbif") \%>\% size(0)
index("gbif") \%>\% size(300)

index("gbif") \%>\% size(1)
index("gbif") \%>\% size(1) \%>\% from(1)
}
}

