% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec.R, R/utils.R
\name{exec}
\alias{exec}
\alias{fields}
\alias{utils}
\title{Execute Elasticsearch query}
\usage{
exec(.obj, query, ...)

fields(x, ...)
}
\arguments{
\item{.obj}{Input}

\item{query}{Query statement}

\item{...}{Further args passed on to \code{\link[jsonlite]{toJSON}}}

\item{x}{Input}

\item{pretty}{Pretty print}

\item{y}{Input}

\item{.dots}{Input}
}
\description{
Execute Elasticsearch query

elastic DSL utilities
}
\details{
Various utilities.
\itemize{
 \item n - Get a count of number of documents found
}
}
\examples{
\dontrun{
# sort
x <- index("gbif") \%>\% filter()
x <- x \%>\% sort(county)
x \%>\% glimpse
x \%>\% exec

x <- index("gbif") \%>\% filter()
x <- x \%>\% sort(asc(county))
x \%>\% glimpse
x \%>\% exec

x <- index("gbif") \%>\% filter()
x <- x \%>\% sort(asc(county), desc(country))
x \%>\% glimpse
x \%>\% exec
}
}

