% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{utils}
\alias{exec}
\alias{fields}
\alias{glimpse}
\alias{n}
\alias{size}
\alias{sort}
\alias{sort_}
\alias{utils}
\title{elastic DSL utilities}
\usage{
n(x)

glimpse(x, pretty = TRUE, ...)

exec(.obj, query, ...)

fields(x, ...)

size(x, y)

sort(.obj = list(), ...)

sort_(.obj = list(), ..., .dots)
}
\arguments{
\item{x}{Input}

\item{...}{Further args passed on to \code{\link[jsonlite]{toJSON}}}
}
\description{
elastic DSL utilities
}
\details{
Various utilities.
\itemize{
 \item n - Get a count of number of documents found
}
}
\examples{
\dontrun{
# sort
x <- index("gbif") \%>\% filter()
x <- x \%>\% sort(county)
x \%>\% glimpse
x \%>\% exec

x <- index("gbif") \%>\% filter()
x <- x \%>\% sort(asc(county))
x \%>\% glimpse
x \%>\% exec

x <- index("gbif") \%>\% filter()
x <- x \%>\% sort(asc(county), desc(country))
x \%>\% glimpse
x \%>\% exec
}
}

