% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{utils}
\alias{exec}
\alias{explain}
\alias{n}
\alias{utils}
\title{elastic DSL utilities}
\usage{
n(x)

explain(x, pretty = TRUE, ...)

exec(.obj, query)
}
\arguments{
\item{x}{Input}

\item{...}{Further args passed on to \code{\link[jsonlite]{toJSON}}}
}
\description{
elastic DSL utilities
}
\details{
Various utilities.
\itemize{
 \item n - Get a count of number of documents found
}
}

